fix: refactoring items folder structure
