+++ be_task_ca/app.py
@@ -1,6 +1,7 @@
 from fastapi import FastAPI, Request, Response
 from .user.api import user_router
-from .item.api import item_router
+# from .item.api import item_router
+from .item.api.api import item_router
 from be_task_ca.user import model
 
 from .database import SessionLocal, engine
