+++ be_task_ca/item/entities/item.py
@@ -1,0 +1,13 @@
+from dataclasses import dataclass, field
+from uuid import UUID, uuid4
+
+
+@dataclass
+class Item:
+    """???"""
+
+    name: str
+    description: str
+    price: float
+    quantity: int
+    id: UUID = field(default_factory=uuid4)
