+++ be_task_ca/user/api.py
@@ -4,7 +4,7 @@ from sqlalchemy.orm import Session
 
 from ..common import get_db
 
-from .usecases import add_item_to_cart, create_user, list_items_in_cart
+# from .usecases import add_item_to_cart, create_user, list_items_in_cart
 
 from .schema import AddToCartRequest, CreateUserRequest
 
@@ -15,18 +15,18 @@ user_router = APIRouter(
 )
 
 
-@user_router.post("/")
-async def post_customer(user: CreateUserRequest, db: Session = Depends(get_db)):
-    return create_user(user, db)
+# @user_router.post("/")
+# async def post_customer(user: CreateUserRequest, db: Session = Depends(get_db)):
+#     return create_user(user, db)
 
 
-@user_router.post("/{user_id}/cart")
-async def post_cart(
-    user_id: UUID, cart_item: AddToCartRequest, db: Session = Depends(get_db)
-):
-    return add_item_to_cart(user_id, cart_item, db)
+# @user_router.post("/{user_id}/cart")
+# async def post_cart(
+#     user_id: UUID, cart_item: AddToCartRequest, db: Session = Depends(get_db)
+# ):
+#     return add_item_to_cart(user_id, cart_item, db)
 
 
-@user_router.get("/{user_id}/cart")
-async def get_cart(user_id: UUID, db: Session = Depends(get_db)):
-    return list_items_in_cart(user_id, db)
+# @user_router.get("/{user_id}/cart")
+# async def get_cart(user_id: UUID, db: Session = Depends(get_db)):
+#     return list_items_in_cart(user_id, db)
